[DP-149] refactor: 파일 분리 #112
