+++ src/app/chat/(default)/page.tsx
@@ -1,5 +1,10 @@
-import ChatList from "@feature/chat/components/sections/list/ChatList";
+import { Suspense } from "react";
+import ChatListPageContent from "@/feature/chat/components/pages/ChatListPageContent";
 
-export default function ChatListPage() {
-  return <ChatList />;
+export default function Page() {
+  return (
+    <Suspense fallback={<div>로딩 중...</div>}>
+      <ChatListPageContent />
+    </Suspense>
+  );
 }
