+++ src/feature/chat/components/pages/ChatListPageContent.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import ChatList from "@feature/chat/components/sections/list/ChatList";
+
+export default function ChatListPage() {
+  return <ChatList />;
+}
